+++ zephyr/CMakeLists.txt
@@ -388,7 +388,7 @@ zephyr_library_sources(
 	# SOF core infrastructure - runs on top of Zephyr
 	${SOF_SRC_PATH}/init/init.c
 	${SOF_SRC_PATH}/init/ext_manifest.c
-	${SOF_SRC_PATH}/schedule/timer_domain.c
+	${SOF_SRC_PATH}/schedule/zephyr_domain.c
 	${SOF_SRC_PATH}/schedule/schedule.c
 	${SOF_SRC_PATH}/schedule/dma_single_chan_domain.c
 	${SOF_SRC_PATH}/schedule/dma_multi_chan_domain.c
